Auth-service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
openai==1.54.4
packaging==24.2
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.6.0
pymongo==4.10.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.20.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
